+++ src/main/java/homeTry/exerciseList/dto/ExerciseDto.java
@@ -5,16 +5,16 @@ import homeTry.exerciseList.model.entity.ExerciseHistory;
 import homeTry.exerciseList.model.entity.ExerciseTime;
 import java.time.Duration;
 
-public record ExerciseResponse(
+public record ExerciseDto(
     Long exerciseId,
     String exerciseName,
     Duration exerciseTime,
     boolean isActive
 ) {
 
-    public static ExerciseResponse of(Exercise exercise, Duration exerciseTime,
+    public static ExerciseDto of(Exercise exercise, Duration exerciseTime,
         boolean isActive) {
-        return new ExerciseResponse(
+        return new ExerciseDto(
             exercise.getExerciseId(),
             exercise.getExerciseName(),
             exerciseTime,
@@ -22,11 +22,11 @@ public record ExerciseResponse(
         );
     }
 
-    public static ExerciseResponse from(ExerciseHistory history) {
+    public static ExerciseDto from(ExerciseHistory history) {
         return of(history.getExercise(), history.getExerciseHistoryTime(), false);
     }
 
-    public static ExerciseResponse from(ExerciseTime exerciseTime) {
+    public static ExerciseDto from(ExerciseTime exerciseTime) {
         return of(exerciseTime.getExercise(), exerciseTime.getExerciseTime(),
             exerciseTime.isActive());
     }
