+++ src/main/java/homeTry/exerciseList/dto/request/ExerciseRequest.java
@@ -1,4 +1,4 @@
-package homeTry.exerciseList.dto;
+package homeTry.exerciseList.dto.request;
 
 public record ExerciseRequest(
     String exerciseName) {
